requirements.txt
atlassian-python-api
#py-gitea
git+https://github.com/Langenfeld/py-gitea.git
PyGithub
